Add ViewEntryActivity
ViewEntryActivity supports a tabbed view which is divided in two tabs, one
labeled "overview" and the other labeled "jisho". The jisho tab displays the
jisho search results for the relevant word, whereas the overview is going to
display info gathered from the jisho API and from the app's own database. The
overview tab currently has no content.

Also, I added an edit icon on the toolbar which is supposed to bring the
activity into edit mode. I've also added a floating action button with the same
icon, which should do the same thing, but it does not seem to display. I need to
figure out why.