Add option to export list as CSV
This is just to have a simple way to back up data in case I need to perform
changes to the database. I should be able to write a method which allows me to
automatically recover all or most of my lost data this way. I might also use the
CSV strings generated as a basis for a better database upgrade/downgrade policy.