Fix #35 - edit button doesn't crash
The fix was to use a workaround found here:
https://stackoverflow.com/a/18611036/4498826

I made a method called getFragment and use it whenever I'm retrieving a
fragment. It turns out that the only place I actually need to null check it,
though, is when using setEditMode in ViewEntryActivity. The important part is to
not call the setEditable method outside of the Fragment - internal calls appear
to be safe.