Close #38 - add data to DB from CSV
Capabilities of adding data to CSV from database has been added. It is now
possible to add CSV data from the named lists screen toolbar action item.
The insertion algorithm is stupid, and only checks if format is correct before
trying to insert the entry - does not validate fields either, except for the
integer fields - needs in particular to validate the reading field and the kanji
field.

It is currently not possible to add CSV from a file. TODO: Continue working on
issue #36 by adding a file selector and allow the user to select a file using
that when clicking the EditText field.